requirements.txt
cholespy==1.0.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
gpytoolbox==0.2.0
imageio==2.34.1
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
libigl==2.5.1
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-image==0.22.0
scipy==1.13.0
six==1.16.0
sympy==1.12
tifffile==2024.5.10
torch==2.2.2
tqdm==4.66.4
typing_extensions==4.11.0
zipp==3.18.2
